+++ packages/ragbits-core/src/ragbits/core/llms/base.py
@@ -1,9 +1,11 @@
 import enum
+import json
 from abc import ABC, abstractmethod
-from collections.abc import AsyncGenerator
-from typing import ClassVar, Generic, TypeVar, cast, overload
+from collections.abc import AsyncGenerator, Callable
+from typing import ClassVar, Generic, TypeVar, overload
 
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict, field_validator
+from typing_extensions import deprecated
 
 from ragbits.core import llms
 from ragbits.core.audit.traces import trace
@@ -16,8 +18,10 @@ from ragbits.core.prompt.base import (
     SimplePrompt,
 )
 from ragbits.core.utils.config_handling import ConfigurableComponent
+from ragbits.core.utils.function_schema import convert_function_to_function_schema
 
 LLMClientOptionsT = TypeVar("LLMClientOptionsT", bound=Options)
+Tool = Callable | dict
 
 
 class LLMType(enum.Enum):
@@ -30,13 +34,32 @@ class LLMType(enum.Enum):
     STRUCTURED_OUTPUT = "structured_output"
 
 
+class ToolCall(BaseModel):
+    """
+    A schema of tool call data
+    """
+
+    name: str
+    arguments: dict
+    model_config = ConfigDict(extra="allow")
+
+    @field_validator("arguments", mode="before")
+    def parse_tool_arguments(cls, tool_arguments: str) -> dict:
+        """
+        Parser for converting tool arguments from string representation to dict
+        """
+        pased_arguments = json.loads(tool_arguments)
+        return pased_arguments
+
+
 class LLMResponseWithMetadata(BaseModel, Generic[PromptOutputT]):
     """
     A schema of output with metadata
     """
 
     content: PromptOutputT
-    metadata: dict
+    metadata: dict = {}
+    tool_calls: list[ToolCall] | None = None
 
 
 class LLM(ConfigurableComponent[LLMClientOptionsT], ABC):
@@ -90,6 +113,7 @@ class LLM(ConfigurableComponent[LLMClientOptionsT], ABC):
         """
         raise NotImplementedError("Token id lookup is not supported by this model")
 
+    @deprecated("Use generate_with_metadata() instead")
     async def generate_raw(
         self,
         prompt: BasePrompt | str | ChatFormat,
@@ -124,41 +148,46 @@ class LLM(ConfigurableComponent[LLMClientOptionsT], ABC):
     @overload
     async def generate(
         self,
-        prompt: BasePromptWithParser[PromptOutputT],
+        prompt: BasePrompt | BasePromptWithParser[PromptOutputT],
         *,
+        tools: None = None,
         options: LLMClientOptionsT | None = None,
     ) -> PromptOutputT: ...
 
     @overload
     async def generate(
         self,
-        prompt: BasePrompt,
+        prompt: BasePrompt | BasePromptWithParser[PromptOutputT],
         *,
+        tools: list[Tool],
         options: LLMClientOptionsT | None = None,
-    ) -> PromptOutputT: ...
+    ) -> PromptOutputT | list[ToolCall]: ...
 
     @overload
     async def generate(
         self,
-        prompt: str,
+        prompt: str | ChatFormat,
         *,
+        tools: None = None,
         options: LLMClientOptionsT | None = None,
     ) -> str: ...
 
     @overload
     async def generate(
         self,
-        prompt: ChatFormat,
+        prompt: str | ChatFormat,
         *,
+        tools: list[Tool],
         options: LLMClientOptionsT | None = None,
-    ) -> str: ...
+    ) -> str | list[ToolCall]: ...
 
     async def generate(
         self,
-        prompt: BasePrompt | str | ChatFormat,
+        prompt: str | ChatFormat | BasePrompt | BasePromptWithParser[PromptOutputT],
         *,
+        tools: list[Tool] | None = None,
         options: LLMClientOptionsT | None = None,
-    ) -> PromptOutputT:
+    ) -> str | PromptOutputT | list[ToolCall]:
         """
         Prepares and sends a prompt to the LLM and returns the parsed response.
 
@@ -167,98 +196,104 @@ class LLM(ConfigurableComponent[LLMClientOptionsT], ABC):
                 - BasePrompt instance: Formatted prompt template with conversation
                 - str: Simple text prompt that will be sent as a user message
                 - ChatFormat: List of message dictionaries in OpenAI chat format
+            tools: Functions to be used as tools by the LLM.
             options: Options to use for the LLM client.
 
         Returns:
-            Parsed response from LLM.
+            Parsed response from LLM or list of tool calls.
         """
-        with trace(model_name=self.model_name, prompt=prompt, options=repr(options)) as outputs:
-            raw_response = await self.generate_raw(prompt, options=options)
-            if isinstance(prompt, BasePromptWithParser):
-                response = await prompt.parse_response(raw_response["response"])
-            else:
-                response = cast(PromptOutputT, raw_response["response"])
-            raw_response["response"] = response
-            outputs.response = raw_response
-
-        return response
+        response = await self.generate_with_metadata(prompt, tools=tools, options=options)
+        return response.tool_calls if tools and response.tool_calls else response.content
 
     @overload
     async def generate_with_metadata(
         self,
-        prompt: BasePromptWithParser[PromptOutputT],
+        prompt: BasePrompt | BasePromptWithParser[PromptOutputT],
         *,
+        tools: list[Tool] | None = None,
         options: LLMClientOptionsT | None = None,
     ) -> LLMResponseWithMetadata[PromptOutputT]: ...
 
     @overload
     async def generate_with_metadata(
         self,
-        prompt: BasePrompt,
+        prompt: str | ChatFormat,
         *,
+        tools: list[Tool] | None = None,
         options: LLMClientOptionsT | None = None,
-    ) -> LLMResponseWithMetadata[PromptOutputT]: ...
-
-    @overload
-    async def generate_with_metadata(
-        self,
-        prompt: str,
-        *,
-        options: LLMClientOptionsT | None = None,
-    ) -> LLMResponseWithMetadata[PromptOutputT]: ...
-
-    @overload
-    @overload
-    async def generate_with_metadata(
-        self,
-        prompt: ChatFormat,
-        *,
-        options: LLMClientOptionsT | None = None,
-    ) -> LLMResponseWithMetadata[PromptOutputT]: ...
+    ) -> LLMResponseWithMetadata[str]: ...
 
     async def generate_with_metadata(
         self,
-        prompt: BasePrompt | str | ChatFormat,
+        prompt: str | ChatFormat | BasePrompt | BasePromptWithParser[PromptOutputT],
         *,
+        tools: list[Tool] | None = None,
         options: LLMClientOptionsT | None = None,
-    ) -> LLMResponseWithMetadata[PromptOutputT]:
+    ) -> LLMResponseWithMetadata[str] | LLMResponseWithMetadata[PromptOutputT]:
         """
         Prepares and sends a prompt to the LLM and returns response parsed to the
         output type of the prompt (if available).
 
         Args:
             prompt: Formatted prompt template with conversation and optional response parsing configuration.
+            tools: Functions to be used as tools by the LLM.
             options: Options to use for the LLM client.
 
         Returns:
-            Text response from LLM with metadata.
+            Text response from LLM with metadata and list of tool calls.
         """
-        with trace(model_name=self.model_name, prompt=prompt, options=repr(options)) as outputs:
-            response = await self.generate_raw(prompt, options=options)
-            content = response.pop("response")
-            if isinstance(prompt, BasePromptWithParser):
+        with trace(name="generate", model_name=self.model_name, prompt=prompt, options=repr(options)) as outputs:
+            merged_options = (self.default_options | options) if options else self.default_options
+
+            if isinstance(prompt, str | list):
+                prompt = SimplePrompt(prompt)
+
+            parsed_tools = (
+                [convert_function_to_function_schema(tool) if callable(tool) else tool for tool in tools]
+                if tools
+                else None
+            )
+            response = await self._call(
+                prompt=prompt,
+                options=merged_options,
+                json_mode=prompt.json_mode,
+                output_schema=prompt.output_schema(),
+                tools=parsed_tools,
+            )
+
+            tool_calls = (
+                [ToolCall.model_validate(tool_call) for tool_call in _tool_calls]
+                if (_tool_calls := response.pop("tool_calls", None)) and tools
+                else None
+            )
+            content = response.pop("response", None)
+            if isinstance(prompt, BasePromptWithParser) and content:
                 content = await prompt.parse_response(content)
+
             outputs.response = LLMResponseWithMetadata[type(content)](  # type: ignore
                 content=content,
+                tool_calls=tool_calls,
                 metadata=response,
             )
-        return outputs.response
+            return outputs.response
 
     async def generate_streaming(
         self,
-        prompt: BasePrompt | str | ChatFormat,
+        prompt: str | ChatFormat | BasePrompt,
         *,
+        tools: list[Tool] | None = None,
         options: LLMClientOptionsT | None = None,
-    ) -> AsyncGenerator[str, None]:
+    ) -> AsyncGenerator[str | ToolCall, None]:
         """
         Prepares and sends a prompt to the LLM and streams the results.
 
         Args:
             prompt: Formatted prompt template with conversation.
-            options: Options to use for the LLM client.
+            tools: Functions to be used as tools by the LLM.
+            options: Options to use for the LLM.
 
         Returns:
-            Response stream from LLM.
+            Response stream from LLM or list of tool calls.
         """
         with trace(model_name=self.model_name, prompt=prompt, options=repr(options)) as outputs:
             merged_options = (self.default_options | options) if options else self.default_options
@@ -266,16 +301,36 @@ class LLM(ConfigurableComponent[LLMClientOptionsT], ABC):
             if isinstance(prompt, str | list):
                 prompt = SimplePrompt(prompt)
 
+            parsed_tools = (
+                [convert_function_to_function_schema(tool) if callable(tool) else tool for tool in tools]
+                if tools
+                else None
+            )
             response = await self._call_streaming(
                 prompt=prompt,
                 options=merged_options,
                 json_mode=prompt.json_mode,
                 output_schema=prompt.output_schema(),
+                tools=parsed_tools,
             )
-            outputs.response = ""
-            async for text in response:
-                outputs.response += text
-                yield text
+
+            content = ""
+            tool_calls = []
+            async for chunk in response:
+                if text := chunk.get("response"):
+                    content += text
+                    yield text
+
+                if tools and (_tool_calls := chunk.get("tool_calls")):
+                    for tool_call in _tool_calls:
+                        parsed_tool_call = ToolCall.model_validate(tool_call)
+                        tool_calls.append(parsed_tool_call)
+                        yield parsed_tool_call
+
+            outputs.response = LLMResponseWithMetadata[type(content or None)](  # type: ignore
+                content=content or None,
+                tool_calls=tool_calls or None,
+            )
 
     @abstractmethod
     async def _call(
@@ -284,15 +339,17 @@ class LLM(ConfigurableComponent[LLMClientOptionsT], ABC):
         options: LLMClientOptionsT,
         json_mode: bool = False,
         output_schema: type[BaseModel] | dict | None = None,
+        tools: list[dict] | None = None,
     ) -> dict:
         """
         Calls LLM inference API.
 
         Args:
             prompt: Formatted prompt template with conversation.
-            options: Additional settings used by LLM.
+            options: Additional settings used by the LLM.
             json_mode: Force the response to be in JSON format.
             output_schema: Schema for structured response (either Pydantic model or a JSON schema).
+            tools: Functions to be used as tools by the LLM.
 
         Returns:
             Response dict from LLM.
@@ -305,16 +362,18 @@ class LLM(ConfigurableComponent[LLMClientOptionsT], ABC):
         options: LLMClientOptionsT,
         json_mode: bool = False,
         output_schema: type[BaseModel] | dict | None = None,
-    ) -> AsyncGenerator[str, None]:
+        tools: list[dict] | None = None,
+    ) -> AsyncGenerator[dict, None]:
         """
         Calls LLM inference API with output streaming.
 
         Args:
             prompt: Formatted prompt template with conversation.
-            options: Additional settings used by LLM.
+            options: Additional settings used by the LLM.
             json_mode: Force the response to be in JSON format.
             output_schema: Schema for structured response (either Pydantic model or a JSON schema).
+            tools: Functions to be used as tools by the LLM.
 
         Returns:
-            Response stream from LLM.
+            Response dict stream from LLM.
         """