+++ packages/ragbits-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add support for async component factories (#608)
+- Add function calling for LLMs (#533)
 
 ## 1.0.0 (2025-06-04)